requirement.txt
PyQt6
requests
python-dotenv